refactor date timezone parsing
